[Job Launcher] Move cron jobs into cron jobs controller (#1514)
* Move cron jobs into cron jobs controller

* Move all cron methods to cron service

* Remove cron jobs from vercel.json
