+++ packages/apps/job-launcher/server/src/modules/cron-job/cron-job.service.spec.ts
@@ -6,12 +6,77 @@ import { CronJobService } from './cron-job.service';
 import { CronJobRepository } from './cron-job.repository';
 import { CronJobEntity } from './cron-job.entity';
 import { createMock } from '@golevelup/ts-jest';
+import { JobEntity } from '../job/job.entity';
+import { JobRequestType, JobStatus } from '../../common/enums/job';
+import {
+  MOCK_ADDRESS,
+  MOCK_EXCHANGE_ORACLE_ADDRESS,
+  MOCK_EXCHANGE_ORACLE_WEBHOOK_URL,
+  MOCK_FILE_HASH,
+  MOCK_FILE_URL,
+  MOCK_MAX_RETRY_COUNT,
+  MOCK_TRANSACTION_HASH,
+} from '../../../test/constants';
+import {
+  ChainId,
+  Encryption,
+  EscrowClient,
+  KVStoreClient,
+} from '@human-protocol/sdk';
+import { JobService } from '../job/job.service';
+import { DeepPartial } from 'typeorm';
+import { CvatManifestDto } from '../job/job.dto';
+import { WebhookService } from '../webhook/webhook.service';
+import { StorageService } from '../storage/storage.service';
+import { Web3Service } from '../web3/web3.service';
+import { PaymentService } from '../payment/payment.service';
+import { JobRepository } from '../job/job.repository';
+import { PaymentRepository } from '../payment/payment.repository';
+import { ConfigService } from '@nestjs/config';
+import { RoutingProtocolService } from '../job/routing-protocol.service';
+import { WebhookEntity } from '../webhook/webhook.entity';
+import { WebhookStatus } from '../../common/enums/webhook';
+import { WebhookRepository } from '../webhook/webhook.repository';
+import { HttpService } from '@nestjs/axios';
+
+jest.mock('@human-protocol/sdk', () => ({
+  ...jest.requireActual('@human-protocol/sdk'),
+  EscrowClient: {
+    build: jest.fn().mockImplementation(() => ({
+      createEscrow: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      setup: jest.fn().mockResolvedValue(null),
+      fund: jest.fn().mockResolvedValue(null),
+    })),
+  },
+  KVStoreClient: {
+    build: jest.fn().mockImplementation(() => ({
+      get: jest.fn(),
+    })),
+  },
+}));
 
 describe('CronJobService', () => {
-  let service: CronJobService;
-  let repository: CronJobRepository;
+  let service: CronJobService,
+    repository: CronJobRepository,
+    webhookService: WebhookService,
+    webhookRepository: WebhookRepository,
+    storageService: StorageService,
+    jobService: JobService;
+
+  const signerMock = {
+    address: MOCK_ADDRESS,
+    getNetwork: jest.fn().mockResolvedValue({ chainId: 1 }),
+  };
 
   beforeEach(async () => {
+    const mockConfigService: Partial<ConfigService> = {
+      get: jest.fn((key: string) => {
+        switch (key) {
+          case 'MAX_RETRY_COUNT':
+            return MOCK_MAX_RETRY_COUNT;
+        }
+      }),
+    };
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         CronJobService,
@@ -19,11 +84,42 @@ describe('CronJobService', () => {
           provide: CronJobRepository,
           useValue: createMock<CronJobRepository>(),
         },
+        {
+          provide: Web3Service,
+          useValue: {
+            getSigner: jest.fn().mockReturnValue(signerMock),
+            validateChainId: jest.fn().mockReturnValue(new Error()),
+          },
+        },
+        JobService,
+        WebhookService,
+        Encryption,
+        { provide: JobRepository, useValue: createMock<JobRepository>() },
+        {
+          provide: PaymentRepository,
+          useValue: createMock<PaymentRepository>(),
+        },
+        { provide: StorageService, useValue: createMock<StorageService>() },
+        { provide: PaymentService, useValue: createMock<PaymentService>() },
+        { provide: ConfigService, useValue: mockConfigService },
+        {
+          provide: RoutingProtocolService,
+          useValue: createMock<RoutingProtocolService>(),
+        },
+        {
+          provide: WebhookRepository,
+          useValue: createMock<WebhookRepository>(),
+        },
+        { provide: HttpService, useValue: createMock<HttpService>() },
       ],
     }).compile();
 
     service = module.get<CronJobService>(CronJobService);
+    jobService = module.get<JobService>(JobService);
     repository = module.get<CronJobRepository>(CronJobRepository);
+    webhookService = module.get<WebhookService>(WebhookService);
+    webhookRepository = module.get<WebhookRepository>(WebhookRepository);
+    storageService = module.get<StorageService>(StorageService);
   });
 
   describe('startCronJob', () => {
@@ -167,6 +263,643 @@ describe('CronJobService', () => {
 
       await expect(service.completeCronJob(cronJobEntity)).rejects.toThrow();
       expect(saveSpy).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('createEscrowCronJob', () => {
+    let createEscrowMock: any;
+    let cronJobEntityMock: Partial<CronJobEntity>;
+    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
+
+    beforeEach(() => {
+      cronJobEntityMock = {
+        cronJobType: CronJobType.CreateEscrow,
+        startedAt: new Date(),
+        save: jest.fn(),
+      };
+
+      jobEntityMock1 = {
+        status: JobStatus.PAID,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jobEntityMock2 = {
+        status: JobStatus.PAID,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jest
+        .spyOn(jobService, 'findJobByStatus')
+        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
+
+      createEscrowMock = jest.spyOn(jobService, 'createEscrow');
+      createEscrowMock.mockResolvedValue(true);
+
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValue(false);
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should not run if the cron job is already running', async () => {
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValueOnce(true);
+
+      await service.createEscrowCronJob();
+
+      expect(createEscrowMock).not.toHaveBeenCalled();
+    });
+
+    it('should create cron job entity on database to lock', async () => {
+      jest
+        .spyOn(service, 'startCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.createEscrowCronJob();
+
+      expect(service.startCronJob).toHaveBeenCalledWith(
+        CronJobType.CreateEscrow,
+      );
+    });
+
+    it('should run createEscrow for all of the jobs with status PAID', async () => {
+      await service.createEscrowCronJob();
+
+      expect(createEscrowMock).toHaveBeenCalledTimes(2);
+    });
+
+    it('should increase retriesCount by 1, if the job creation fails', async () => {
+      createEscrowMock.mockRejectedValueOnce(new Error('creation failed'));
+
+      await service.createEscrowCronJob();
+
+      expect(createEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.retriesCount).toBe(2);
+      expect(jobEntityMock2.retriesCount).toBe(1);
+    });
+
+    it('should mark job as failed if the job creation fails more than max retries count', async () => {
+      createEscrowMock.mockRejectedValueOnce(new Error('creation failed'));
+      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await service.createEscrowCronJob();
+
+      expect(createEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
+      expect(jobEntityMock2.status).toBe(JobStatus.PAID);
+    });
+
+    it('should complete the cron job entity on database to unlock', async () => {
+      jest
+        .spyOn(service, 'completeCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.createEscrowCronJob();
+
+      expect(service.completeCronJob).toHaveBeenCalledWith(
+        CronJobType.CreateEscrow,
+      );
+    });
+  });
+
+  describe('setupEscrowCronJob', () => {
+    let setupEscrowMock: any;
+    let cronJobEntityMock: Partial<CronJobEntity>;
+    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
+
+    beforeEach(() => {
+      cronJobEntityMock = {
+        cronJobType: CronJobType.SetupEscrow,
+        createdAt: new Date(),
+        save: jest.fn(),
+      };
+
+      jobEntityMock1 = {
+        status: JobStatus.CREATED,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jobEntityMock2 = {
+        status: JobStatus.CREATED,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jest
+        .spyOn(jobService, 'findJobByStatus')
+        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
+
+      setupEscrowMock = jest.spyOn(jobService, 'setupEscrow');
+      setupEscrowMock.mockResolvedValue(true);
+
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValue(false);
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should not run if the cron job is already running', async () => {
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValueOnce(true);
+
+      await service.setupEscrowCronJob();
+
+      expect(setupEscrowMock).not.toHaveBeenCalled();
+    });
+
+    it('should create cron job entity on database to lock', async () => {
+      jest
+        .spyOn(service, 'startCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.setupEscrowCronJob();
+
+      expect(service.startCronJob).toHaveBeenCalledWith(
+        CronJobType.SetupEscrow,
+      );
+    });
+
+    it('should run setupEscrow for all of the jobs with status LAUNCHING', async () => {
+      await service.setupEscrowCronJob();
+
+      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
+    });
+
+    it('should increase retriesCount by 1, if the job setup fails', async () => {
+      setupEscrowMock.mockRejectedValueOnce(new Error('setup failed'));
+
+      await service.setupEscrowCronJob();
+
+      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.retriesCount).toBe(2);
+      expect(jobEntityMock2.retriesCount).toBe(1);
+    });
+
+    it('should mark job as failed if the job setup fails more than max retries count', async () => {
+      setupEscrowMock.mockRejectedValueOnce(new Error('setup failed'));
+      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await service.setupEscrowCronJob();
+
+      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
+      expect(jobEntityMock2.status).toBe(JobStatus.CREATED);
+    });
+
+    it('should complete the cron job entity on database to unlock', async () => {
+      jest
+        .spyOn(service, 'completeCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.setupEscrowCronJob();
+
+      expect(service.completeCronJob).toHaveBeenCalledWith(
+        CronJobType.SetupEscrow,
+      );
+    });
+  });
+
+  describe('fundEscrowCronJob', () => {
+    let fundEscrowMock: any;
+    let cronJobEntityMock: Partial<CronJobEntity>;
+    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
+    let createWebhookMock: any;
+
+    beforeEach(() => {
+      cronJobEntityMock = {
+        cronJobType: CronJobType.FundEscrow,
+        createdAt: new Date(),
+        save: jest.fn(),
+      };
+
+      jobEntityMock1 = {
+        status: JobStatus.SET_UP,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jobEntityMock2 = {
+        status: JobStatus.SET_UP,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        manifestHash: MOCK_FILE_HASH,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        retriesCount: 1,
+        save: jest.fn(),
+      };
+
+      jest
+        .spyOn(jobService, 'findJobByStatus')
+        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
+
+      fundEscrowMock = jest.spyOn(jobService, 'fundEscrow');
+      fundEscrowMock.mockResolvedValue(true);
+
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValue(false);
+
+      createWebhookMock = jest.spyOn(webhookService, 'createWebhook');
+
+      const cvatManifestMock: DeepPartial<CvatManifestDto> = {
+        data: {
+          data_url: MOCK_FILE_URL,
+        },
+        annotation: {
+          type: JobRequestType.IMAGE_POINTS,
+        },
+      };
+      jest
+        .spyOn(storageService, 'download')
+        .mockResolvedValue(cvatManifestMock);
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should not run if the cron job is already running', async () => {
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValueOnce(true);
+
+      await service.fundEscrowCronJob();
+
+      expect(fundEscrowMock).not.toHaveBeenCalled();
+    });
+
+    it('should create cron job entity on database to lock', async () => {
+      jest
+        .spyOn(service, 'startCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.fundEscrowCronJob();
+
+      expect(service.startCronJob).toHaveBeenCalledWith(CronJobType.FundEscrow);
+    });
+
+    it('should run fundEscrow for all of the jobs with status FUNDING, and trigger webhook', async () => {
+      await service.fundEscrowCronJob();
+
+      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
+      expect(createWebhookMock).toHaveBeenCalledTimes(2);
+    });
+
+    it('should increase retriesCount by 1, if the job fund fails', async () => {
+      fundEscrowMock.mockRejectedValueOnce(new Error('fund failed'));
+
+      await service.fundEscrowCronJob();
+
+      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.retriesCount).toBe(2);
+      expect(jobEntityMock2.retriesCount).toBe(1);
+
+      expect(createWebhookMock).toHaveBeenCalledTimes(1);
+    });
+
+    it('should mark job as failed if the job fund fails more than max retries count', async () => {
+      fundEscrowMock.mockRejectedValueOnce(new Error('fund failed'));
+      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await service.fundEscrowCronJob();
+
+      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
+      expect(jobEntityMock2.status).toBe(JobStatus.SET_UP);
+
+      expect(createWebhookMock).toHaveBeenCalledTimes(1);
+    });
+
+    it('should complete the cron job entity on database to unlock', async () => {
+      jest
+        .spyOn(service, 'completeCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.fundEscrowCronJob();
+
+      expect(service.completeCronJob).toHaveBeenCalledWith(
+        CronJobType.FundEscrow,
+      );
+    });
+  });
+
+  describe('cancelCronJob', () => {
+    let findJobMock: any,
+      cronJobEntityMock: Partial<CronJobEntity>,
+      jobEntityMock1: Partial<JobEntity>,
+      jobEntityMock2: Partial<JobEntity>;
+
+    beforeEach(() => {
+      cronJobEntityMock = {
+        cronJobType: CronJobType.CancelEscrow,
+        createdAt: new Date(),
+        save: jest.fn(),
+      };
+
+      jobEntityMock1 = {
+        status: JobStatus.TO_CANCEL,
+        fundAmount: 100,
+        userId: 1,
+        id: 1,
+        manifestUrl: MOCK_FILE_URL,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        save: jest.fn(),
+        retriesCount: 0,
+      };
+
+      jobEntityMock2 = {
+        status: JobStatus.TO_CANCEL,
+        fundAmount: 100,
+        userId: 1,
+        id: 2,
+        manifestUrl: MOCK_FILE_URL,
+        escrowAddress: MOCK_ADDRESS,
+        chainId: ChainId.LOCALHOST,
+        save: jest.fn(),
+        retriesCount: 0,
+      };
+
+      findJobMock = jest
+        .spyOn(jobService, 'findJobByStatus')
+        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
+
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValue(false);
+
+      jest.spyOn(jobService, 'processEscrowCancellation').mockResolvedValue({
+        txHash: MOCK_TRANSACTION_HASH,
+        amountRefunded: 1n,
+      });
+
+      (EscrowClient.build as any).mockImplementation(() => ({
+        getExchangeOracleAddress: jest
+          .fn()
+          .mockResolvedValue(MOCK_EXCHANGE_ORACLE_ADDRESS),
+      }));
+
+      (KVStoreClient.build as any).mockImplementation(() => ({
+        get: jest.fn().mockResolvedValue(MOCK_EXCHANGE_ORACLE_WEBHOOK_URL),
+      }));
+
+      const manifestMock = {
+        requestType: JobRequestType.FORTUNE,
+      };
+      storageService.download = jest.fn().mockResolvedValue(manifestMock);
+    });
+
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+
+    it('should not run if cron job is already running', async () => {
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValueOnce(true);
+
+      await service.cancelCronJob();
+
+      expect(findJobMock).not.toHaveBeenCalled();
+    });
+
+    it('should create cron job entity on database to lock', async () => {
+      jest
+        .spyOn(service, 'startCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.cancelCronJob();
+
+      expect(service.startCronJob).toHaveBeenCalledWith(
+        CronJobType.CancelEscrow,
+      );
+    });
+
+    it('should cancel all of the jobs with status TO_CANCEL', async () => {
+      jest.spyOn(webhookService, 'createWebhook');
+      const result = await service.cancelCronJob();
+
+      expect(result).toBeTruthy();
+      expect(jobService.processEscrowCancellation).toHaveBeenCalledWith(
+        jobEntityMock1,
+      );
+      expect(jobEntityMock1.save).toHaveBeenCalled();
+      expect(jobService.processEscrowCancellation).toHaveBeenCalledWith(
+        jobEntityMock2,
+      );
+      expect(jobEntityMock2.save).toHaveBeenCalled();
+      expect(webhookService.createWebhook).toHaveBeenCalledTimes(2);
+    });
+
+    it('should not call process escrow cancellation when escrowAddress is not present', async () => {
+      const jobEntityWithoutEscrow = {
+        ...jobEntityMock1,
+        escrowAddress: undefined,
+      };
+
+      jest
+        .spyOn(jobService, 'findJobByStatus')
+        .mockResolvedValueOnce([jobEntityWithoutEscrow as any]);
+      jest
+        .spyOn(jobService, 'processEscrowCancellation')
+        .mockResolvedValueOnce(undefined as any);
+
+      expect(await service.cancelCronJob()).toBe(true);
+      expect(jobService.processEscrowCancellation).toHaveBeenCalledTimes(0);
+    });
+
+    it('should increase retriesCount by 1 if the job cancellation fails', async () => {
+      jest
+        .spyOn(jobService, 'processEscrowCancellation')
+        .mockRejectedValueOnce(new Error('cancellation failed'));
+
+      expect(jobEntityMock1.retriesCount).toBe(0);
+      expect(jobEntityMock2.retriesCount).toBe(0);
+
+      await service.cancelCronJob();
+
+      expect(jobEntityMock1.retriesCount).toBe(1);
+      expect(jobEntityMock2.retriesCount).toBe(0);
+    });
+
+    it('should mark job as failed if the job cancellation fails more than max retries count', async () => {
+      jest
+        .spyOn(jobService, 'processEscrowCancellation')
+        .mockRejectedValueOnce(new Error('cancellation failed'));
+      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await service.cancelCronJob();
+
+      expect(jobService.processEscrowCancellation).toHaveBeenCalledTimes(2);
+      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
+      expect(jobEntityMock2.status).toBe(JobStatus.CANCELED);
+    });
+
+    it('should complete the cron job entity on database to unlock', async () => {
+      jest.spyOn(service, 'completeCronJob').mockResolvedValueOnce({} as any);
+
+      await service.cancelCronJob();
+
+      expect(service.completeCronJob).toHaveBeenCalledWith(
+        CronJobType.CancelEscrow,
+      );
+    });
+  });
+
+  describe('processPendingCronJob', () => {
+    let sendWebhookMock: any;
+    let cronJobEntityMock: Partial<CronJobEntity>;
+    let webhookEntity1: Partial<WebhookEntity>,
+      webhookEntity2: Partial<WebhookEntity>;
+
+    beforeEach(() => {
+      cronJobEntityMock = {
+        cronJobType: CronJobType.ProcessPendingWebhook,
+        startedAt: new Date(),
+        save: jest.fn(),
+      };
+
+      webhookEntity1 = {
+        id: 1,
+        chainId: ChainId.LOCALHOST,
+        escrowAddress: MOCK_ADDRESS,
+        status: WebhookStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      webhookEntity2 = {
+        id: 2,
+        chainId: ChainId.LOCALHOST,
+        escrowAddress: MOCK_ADDRESS,
+        status: WebhookStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      jest
+        .spyOn(webhookRepository, 'find')
+        .mockResolvedValue([webhookEntity1 as any, webhookEntity2 as any]);
+
+      sendWebhookMock = jest.spyOn(webhookService as any, 'sendWebhook');
+      sendWebhookMock.mockResolvedValue(true);
+
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValue(false);
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should not run if cron job is already running', async () => {
+      jest.spyOn(service, 'isCronJobRunning').mockResolvedValueOnce(true);
+
+      const startCronJobMock = jest.spyOn(service, 'startCronJob');
+
+      await service.processPendingWebhooks();
+
+      expect(startCronJobMock).not.toHaveBeenCalled();
+    });
+
+    it('should create cron job entity to lock the process', async () => {
+      jest
+        .spyOn(service, 'startCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.processPendingWebhooks();
+
+      expect(service.startCronJob).toHaveBeenCalledWith(
+        CronJobType.ProcessPendingWebhook,
+      );
+    });
+
+    it('should send webhook for all of the pending webhooks', async () => {
+      await service.processPendingWebhooks();
+
+      expect(sendWebhookMock).toHaveBeenCalledTimes(2);
+      expect(sendWebhookMock).toHaveBeenCalledWith(webhookEntity1);
+      expect(sendWebhookMock).toHaveBeenCalledWith(webhookEntity2);
+
+      expect(webhookRepository.updateOne).toHaveBeenCalledTimes(2);
+      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
+        { id: webhookEntity1.id },
+        { status: WebhookStatus.COMPLETED },
+      );
+      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
+        { id: webhookEntity2.id },
+        { status: WebhookStatus.COMPLETED },
+      );
+    });
+
+    it('should increase retriesCount by 1 if sending webhook fails', async () => {
+      sendWebhookMock.mockRejectedValueOnce(new Error());
+      await service.processPendingWebhooks();
+
+      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
+        { id: webhookEntity1.id },
+        {
+          retriesCount: 1,
+          waitUntil: expect.any(Date),
+        },
+      );
+    });
+
+    it('should mark webhook as failed if retriesCount exceeds threshold', async () => {
+      sendWebhookMock.mockRejectedValueOnce(new Error());
+
+      webhookEntity1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await service.processPendingWebhooks();
+
+      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
+        { id: webhookEntity1.id },
+        { status: WebhookStatus.FAILED },
+      );
+    });
+
+    it('should complete the cron job entity to unlock', async () => {
+      jest
+        .spyOn(service, 'completeCronJob')
+        .mockResolvedValueOnce(cronJobEntityMock as any);
+
+      await service.processPendingWebhooks();
+
+      expect(service.completeCronJob).toHaveBeenCalledWith(
+        cronJobEntityMock as any,
+      );
     });
   });
 });
