+++ packages/apps/job-launcher/server/src/modules/job/job.controller.ts
@@ -186,63 +186,6 @@ export class JobController {
   }
 
   @ApiOperation({
-    summary: 'Launch the cron job to create escrows',
-    description: 'Endpoint to launch the cron job to create escrows.',
-  })
-  @Public()
-  @ApiResponse({
-    status: 200,
-    description: 'Cron job to create escrows launched successfully.',
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found. Could not find the requested content.',
-  })
-  @Get('/cron/create-escrow')
-  public async launchCreateEscrowCronJob(): Promise<void> {
-    await this.jobService.createEscrowCronJob();
-    return;
-  }
-
-  @ApiOperation({
-    summary: 'Launch the cron job to setup escrows',
-    description: 'Endpoint to launch the cron job to setup escrows.',
-  })
-  @Public()
-  @ApiResponse({
-    status: 200,
-    description: 'Cron job to setup escrows launched successfully.',
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found. Could not find the requested content.',
-  })
-  @Get('/cron/setup-escrow')
-  public async launchSetupEscrowCronJob(): Promise<void> {
-    await this.jobService.setupEscrowCronJob();
-    return;
-  }
-
-  @ApiOperation({
-    summary: 'Launch the cron job to fund escrows',
-    description: 'Endpoint to launch the cron job to fund escrows.',
-  })
-  @Public()
-  @ApiResponse({
-    status: 200,
-    description: 'Cron job to fund escrows launched successfully.',
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found. Could not find the requested content.',
-  })
-  @Get('/cron/fund-escrow')
-  public async launchFundEscrowCronJob(): Promise<void> {
-    await this.jobService.fundEscrowCronJob();
-    return;
-  }
-
-  @ApiOperation({
     summary: 'Cancel a job',
     description: 'Endpoint to cancel a specified job.',
   })
@@ -264,21 +207,6 @@ export class JobController {
     @Param() params: JobCancelDto,
   ): Promise<void> {
     await this.jobService.requestToCancelJob(req.user.id, params.id);
-    return;
-  }
-
-  @ApiOperation({
-    summary: 'Cancel a cron job',
-    description: 'Endpoint to cancel a cron job.',
-  })
-  @Public()
-  @ApiResponse({
-    status: 200,
-    description: 'Cron job launched successfully.',
-  })
-  @Get('/cron/cancel')
-  public async cancelCronJob(): Promise<void> {
-    await this.jobService.cancelCronJob();
     return;
   }
