+++ packages/apps/job-launcher/server/src/modules/job/job.service.spec.ts
@@ -49,7 +49,6 @@ import {
   MOCK_CHAIN_ID,
   MOCK_EXCHANGE_ORACLE_ADDRESS,
   MOCK_ORACLE_FEE,
-  MOCK_EXCHANGE_ORACLE_WEBHOOK_URL,
   MOCK_FILE_HASH,
   MOCK_FILE_URL,
   MOCK_HCAPTCHA_ORACLE_ADDRESS,
@@ -109,9 +108,6 @@ import {
 } from '../../common/constants';
 import { WebhookService } from '../webhook/webhook.service';
 import { CronJobService } from '../cron-job/cron-job.service';
-import { CronJobEntity } from '../cron-job/cron-job.entity';
-import { CronJobType } from '../../common/enums/cron-job';
-import { DeepPartial } from 'typeorm';
 import { AWSRegions, StorageProviders } from '../../common/enums/storage';
 
 const rate = 1.5;
@@ -167,9 +163,7 @@ describe('JobService', () => {
     routingProtocolService: RoutingProtocolService,
     web3Service: Web3Service,
     encryption: Encryption,
-    storageService: StorageService,
-    webhookService: WebhookService,
-    cronJobService: CronJobService;
+    storageService: StorageService;
 
   let encrypt = true;
 
@@ -272,8 +266,6 @@ describe('JobService', () => {
     createPaymentMock = jest.spyOn(paymentRepository, 'create');
     web3Service = moduleRef.get<Web3Service>(Web3Service);
     storageService = moduleRef.get<StorageService>(StorageService);
-    webhookService = moduleRef.get<WebhookService>(WebhookService);
-    cronJobService = moduleRef.get<CronJobService>(CronJobService);
 
     storageService.uploadFile = jest.fn().mockResolvedValue({
       url: MOCK_FILE_URL,
@@ -1638,520 +1630,6 @@ describe('JobService', () => {
         jobService.requestToCancelJob(userId, jobId),
       ).rejects.toThrow(
         new ConflictException(ErrorJob.InvalidStatusCancellation),
-      );
-    });
-  });
-
-  describe('createEscrowCronJob', () => {
-    let createEscrowMock: any;
-    let cronJobEntityMock: Partial<CronJobEntity>;
-    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
-
-    beforeEach(() => {
-      cronJobEntityMock = {
-        cronJobType: CronJobType.CreateEscrow,
-        startedAt: new Date(),
-      };
-
-      jobEntityMock1 = {
-        status: JobStatus.PAID,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jobEntityMock2 = {
-        status: JobStatus.PAID,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jest
-        .spyOn(jobRepository, 'find')
-        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
-
-      createEscrowMock = jest.spyOn(jobService, 'createEscrow');
-      createEscrowMock.mockResolvedValue(true);
-
-      jest.spyOn(cronJobService, 'isCronJobRunning').mockResolvedValue(false);
-    });
-
-    afterEach(() => {
-      jest.restoreAllMocks();
-    });
-
-    it('should not run if the cron job is already running', async () => {
-      jest
-        .spyOn(cronJobService, 'isCronJobRunning')
-        .mockResolvedValueOnce(true);
-
-      await jobService.createEscrowCronJob();
-
-      expect(createEscrowMock).not.toHaveBeenCalled();
-    });
-
-    it('should create cron job entity on database to lock', async () => {
-      jest
-        .spyOn(cronJobService, 'startCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.createEscrowCronJob();
-
-      expect(cronJobService.startCronJob).toHaveBeenCalledWith(
-        CronJobType.CreateEscrow,
-      );
-    });
-
-    it('should run createEscrow for all of the jobs with status PAID', async () => {
-      await jobService.createEscrowCronJob();
-
-      expect(createEscrowMock).toHaveBeenCalledTimes(2);
-    });
-
-    it('should increase retriesCount by 1, if the job creation fails', async () => {
-      createEscrowMock.mockRejectedValueOnce(new Error('creation failed'));
-
-      await jobService.createEscrowCronJob();
-
-      expect(createEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.retriesCount).toBe(2);
-      expect(jobEntityMock2.retriesCount).toBe(1);
-    });
-
-    it('should mark job as failed if the job creation fails more than max retries count', async () => {
-      createEscrowMock.mockRejectedValueOnce(new Error('creation failed'));
-      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
-
-      await jobService.createEscrowCronJob();
-
-      expect(createEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
-      expect(jobEntityMock2.status).toBe(JobStatus.PAID);
-    });
-
-    it('should complete the cron job entity on database to unlock', async () => {
-      jest
-        .spyOn(cronJobService, 'completeCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.createEscrowCronJob();
-
-      expect(cronJobService.completeCronJob).toHaveBeenCalledWith(
-        CronJobType.CreateEscrow,
-      );
-    });
-  });
-
-  describe('setupEscrowCronJob', () => {
-    let setupEscrowMock: any;
-    let cronJobEntityMock: Partial<CronJobEntity>;
-    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
-
-    beforeEach(() => {
-      cronJobEntityMock = {
-        cronJobType: CronJobType.SetupEscrow,
-        createdAt: new Date(),
-      };
-
-      jobEntityMock1 = {
-        status: JobStatus.CREATED,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jobEntityMock2 = {
-        status: JobStatus.CREATED,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jest
-        .spyOn(jobRepository, 'find')
-        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
-
-      setupEscrowMock = jest.spyOn(jobService, 'setupEscrow');
-      setupEscrowMock.mockResolvedValue(true);
-
-      jest.spyOn(cronJobService, 'isCronJobRunning').mockResolvedValue(false);
-    });
-
-    afterEach(() => {
-      jest.restoreAllMocks();
-    });
-
-    it('should not run if the cron job is already running', async () => {
-      jest
-        .spyOn(cronJobService, 'isCronJobRunning')
-        .mockResolvedValueOnce(true);
-
-      await jobService.setupEscrowCronJob();
-
-      expect(setupEscrowMock).not.toHaveBeenCalled();
-    });
-
-    it('should create cron job entity on database to lock', async () => {
-      jest
-        .spyOn(cronJobService, 'startCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.setupEscrowCronJob();
-
-      expect(cronJobService.startCronJob).toHaveBeenCalledWith(
-        CronJobType.SetupEscrow,
-      );
-    });
-
-    it('should run setupEscrow for all of the jobs with status LAUNCHING', async () => {
-      await jobService.setupEscrowCronJob();
-
-      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
-    });
-
-    it('should increase retriesCount by 1, if the job setup fails', async () => {
-      setupEscrowMock.mockRejectedValueOnce(new Error('setup failed'));
-
-      await jobService.setupEscrowCronJob();
-
-      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.retriesCount).toBe(2);
-      expect(jobEntityMock2.retriesCount).toBe(1);
-    });
-
-    it('should mark job as failed if the job setup fails more than max retries count', async () => {
-      setupEscrowMock.mockRejectedValueOnce(new Error('setup failed'));
-      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
-
-      await jobService.setupEscrowCronJob();
-
-      expect(setupEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
-      expect(jobEntityMock2.status).toBe(JobStatus.CREATED);
-    });
-
-    it('should complete the cron job entity on database to unlock', async () => {
-      jest
-        .spyOn(cronJobService, 'completeCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.setupEscrowCronJob();
-
-      expect(cronJobService.completeCronJob).toHaveBeenCalledWith(
-        CronJobType.SetupEscrow,
-      );
-    });
-  });
-
-  describe('fundEscrowCronJob', () => {
-    let fundEscrowMock: any;
-    let cronJobEntityMock: Partial<CronJobEntity>;
-    let jobEntityMock1: Partial<JobEntity>, jobEntityMock2: Partial<JobEntity>;
-    let createWebhookMock: any;
-
-    beforeEach(() => {
-      cronJobEntityMock = {
-        cronJobType: CronJobType.FundEscrow,
-        createdAt: new Date(),
-      };
-
-      jobEntityMock1 = {
-        status: JobStatus.SET_UP,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jobEntityMock2 = {
-        status: JobStatus.SET_UP,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        manifestHash: MOCK_FILE_HASH,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        retriesCount: 1,
-        save: jest.fn(),
-      };
-
-      jest
-        .spyOn(jobRepository, 'find')
-        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
-
-      fundEscrowMock = jest.spyOn(jobService, 'fundEscrow');
-      fundEscrowMock.mockResolvedValue(true);
-
-      jest.spyOn(cronJobService, 'isCronJobRunning').mockResolvedValue(false);
-
-      createWebhookMock = jest.spyOn(webhookService, 'createWebhook');
-
-      const cvatManifestMock: DeepPartial<CvatManifestDto> = {
-        data: {
-          data_url: MOCK_FILE_URL,
-        },
-        annotation: {
-          type: JobRequestType.IMAGE_POINTS,
-        },
-      };
-      jest
-        .spyOn(storageService, 'download')
-        .mockResolvedValue(cvatManifestMock);
-    });
-
-    afterEach(() => {
-      jest.restoreAllMocks();
-    });
-
-    it('should not run if the cron job is already running', async () => {
-      jest
-        .spyOn(cronJobService, 'isCronJobRunning')
-        .mockResolvedValueOnce(true);
-
-      await jobService.fundEscrowCronJob();
-
-      expect(fundEscrowMock).not.toHaveBeenCalled();
-    });
-
-    it('should create cron job entity on database to lock', async () => {
-      jest
-        .spyOn(cronJobService, 'startCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.fundEscrowCronJob();
-
-      expect(cronJobService.startCronJob).toHaveBeenCalledWith(
-        CronJobType.FundEscrow,
-      );
-    });
-
-    it('should run fundEscrow for all of the jobs with status FUNDING, and trigger webhook', async () => {
-      await jobService.fundEscrowCronJob();
-
-      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
-      expect(createWebhookMock).toHaveBeenCalledTimes(2);
-    });
-
-    it('should increase retriesCount by 1, if the job fund fails', async () => {
-      fundEscrowMock.mockRejectedValueOnce(new Error('fund failed'));
-
-      await jobService.fundEscrowCronJob();
-
-      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.retriesCount).toBe(2);
-      expect(jobEntityMock2.retriesCount).toBe(1);
-
-      expect(createWebhookMock).toHaveBeenCalledTimes(1);
-    });
-
-    it('should mark job as failed if the job fund fails more than max retries count', async () => {
-      fundEscrowMock.mockRejectedValueOnce(new Error('fund failed'));
-      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
-
-      await jobService.fundEscrowCronJob();
-
-      expect(fundEscrowMock).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
-      expect(jobEntityMock2.status).toBe(JobStatus.SET_UP);
-
-      expect(createWebhookMock).toHaveBeenCalledTimes(1);
-    });
-
-    it('should complete the cron job entity on database to unlock', async () => {
-      jest
-        .spyOn(cronJobService, 'completeCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await jobService.fundEscrowCronJob();
-
-      expect(cronJobService.completeCronJob).toHaveBeenCalledWith(
-        CronJobType.FundEscrow,
-      );
-    });
-  });
-
-  describe('cancelCronJob', () => {
-    let findJobMock: any,
-      jobEntityMock1: Partial<JobEntity>,
-      jobEntityMock2: Partial<JobEntity>;
-
-    beforeEach(() => {
-      jobEntityMock1 = {
-        status: JobStatus.TO_CANCEL,
-        fundAmount: 100,
-        userId: 1,
-        id: 1,
-        manifestUrl: MOCK_FILE_URL,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        save: jest.fn(),
-        retriesCount: 0,
-      };
-
-      jobEntityMock2 = {
-        status: JobStatus.TO_CANCEL,
-        fundAmount: 100,
-        userId: 1,
-        id: 2,
-        manifestUrl: MOCK_FILE_URL,
-        escrowAddress: MOCK_ADDRESS,
-        chainId: ChainId.LOCALHOST,
-        save: jest.fn(),
-        retriesCount: 0,
-      };
-
-      findJobMock = jest
-        .spyOn(jobRepository, 'find')
-        .mockResolvedValue([jobEntityMock1 as any, jobEntityMock2 as any]);
-
-      jest.spyOn(cronJobService, 'isCronJobRunning').mockResolvedValue(false);
-
-      jest.spyOn(jobService, 'processEscrowCancellation').mockResolvedValue({
-        txHash: MOCK_TRANSACTION_HASH,
-        amountRefunded: 1n,
-      });
-
-      (EscrowClient.build as any).mockImplementation(() => ({
-        getExchangeOracleAddress: jest
-          .fn()
-          .mockResolvedValue(MOCK_EXCHANGE_ORACLE_ADDRESS),
-      }));
-
-      (KVStoreClient.build as any).mockImplementation(() => ({
-        get: jest.fn().mockResolvedValue(MOCK_EXCHANGE_ORACLE_WEBHOOK_URL),
-      }));
-
-      const manifestMock = {
-        requestType: JobRequestType.FORTUNE,
-      };
-      storageService.download = jest.fn().mockResolvedValue(manifestMock);
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('should not run if cron job is already running', async () => {
-      jest
-        .spyOn(cronJobService, 'isCronJobRunning')
-        .mockResolvedValueOnce(true);
-
-      await jobService.cancelCronJob();
-
-      expect(findJobMock).not.toHaveBeenCalled();
-    });
-
-    it('should create cron job entity on database to lock', async () => {
-      jest
-        .spyOn(cronJobService, 'startCronJob')
-        .mockResolvedValueOnce({} as any);
-
-      await jobService.cancelCronJob();
-
-      expect(cronJobService.startCronJob).toHaveBeenCalledWith(
-        CronJobType.CancelEscrow,
-      );
-    });
-
-    it('should cancel all of the jobs with status TO_CANCEL', async () => {
-      const result = await jobService.cancelCronJob();
-
-      expect(result).toBeTruthy();
-      expect(jobService.processEscrowCancellation).toHaveBeenCalledWith(
-        jobEntityMock1,
-      );
-      expect(jobEntityMock1.save).toHaveBeenCalled();
-      expect(jobService.processEscrowCancellation).toHaveBeenCalledWith(
-        jobEntityMock2,
-      );
-      expect(jobEntityMock2.save).toHaveBeenCalled();
-      expect(webhookService.createWebhook).toHaveBeenCalledTimes(2);
-    });
-
-    it('should not call process escrow cancellation when escrowAddress is not present', async () => {
-      const jobEntityWithoutEscrow = {
-        ...jobEntityMock1,
-        escrowAddress: undefined,
-      };
-
-      jest
-        .spyOn(jobRepository, 'find')
-        .mockResolvedValueOnce([jobEntityWithoutEscrow as any]);
-      jest
-        .spyOn(jobService, 'processEscrowCancellation')
-        .mockResolvedValueOnce(undefined as any);
-
-      expect(await jobService.cancelCronJob()).toBe(true);
-      expect(jobService.processEscrowCancellation).toHaveBeenCalledTimes(0);
-    });
-
-    it('should increase retriesCount by 1 if the job cancellation fails', async () => {
-      jest
-        .spyOn(jobService, 'processEscrowCancellation')
-        .mockRejectedValueOnce(new Error('cancellation failed'));
-
-      expect(jobEntityMock1.retriesCount).toBe(0);
-      expect(jobEntityMock2.retriesCount).toBe(0);
-
-      await jobService.cancelCronJob();
-
-      expect(jobEntityMock1.retriesCount).toBe(1);
-      expect(jobEntityMock2.retriesCount).toBe(0);
-    });
-
-    it('should mark job as failed if the job cancellation fails more than max retries count', async () => {
-      jest
-        .spyOn(jobService, 'processEscrowCancellation')
-        .mockRejectedValueOnce(new Error('cancellation failed'));
-      jobEntityMock1.retriesCount = MOCK_MAX_RETRY_COUNT;
-
-      await jobService.cancelCronJob();
-
-      expect(jobService.processEscrowCancellation).toHaveBeenCalledTimes(2);
-      expect(jobEntityMock1.status).toBe(JobStatus.FAILED);
-      expect(jobEntityMock2.status).toBe(JobStatus.CANCELED);
-    });
-
-    it('should complete the cron job entity on database to unlock', async () => {
-      jest
-        .spyOn(cronJobService, 'completeCronJob')
-        .mockResolvedValueOnce({} as any);
-
-      await jobService.cancelCronJob();
-
-      expect(cronJobService.completeCronJob).toHaveBeenCalledWith(
-        CronJobType.CancelEscrow,
       );
     });
   });
