--- packages/apps/job-launcher/server/src/modules/webhook/webhook.controller.ts
@@ -1,17 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-import { Public } from '../../common/decorators';
-import { WebhookService } from './webhook.service';
-
-@Public()
-@ApiTags('Webhook')
-@Controller('/webhook')
-export class WebhookController {
-  constructor(private readonly webhookService: WebhookService) {}
-  @Public()
-  @Get('/cron')
-  public async processPendingWebhooks(): Promise<any> {
-    await this.webhookService.processPendingWebhooks();
-    return;
-  }
-}
