+++ packages/apps/job-launcher/server/src/modules/webhook/webhook.service.spec.ts
@@ -23,9 +23,6 @@ import { WebhookRepository } from './webhook.repository';
 import { WebhookService } from './webhook.service';
 import { of } from 'rxjs';
 import { HEADER_SIGNATURE_KEY } from '../../common/constants';
-import { CronJobService } from '../cron-job/cron-job.service';
-import { CronJobEntity } from '../cron-job/cron-job.entity';
-import { CronJobType } from '../../common/enums/cron-job';
 
 jest.mock('@human-protocol/sdk', () => ({
   ...jest.requireActual('@human-protocol/sdk'),
@@ -41,8 +38,7 @@ describe('WebhookService', () => {
   let webhookService: WebhookService,
     webhookRepository: WebhookRepository,
     web3Service: Web3Service,
-    httpService: HttpService,
-    cronJobService: CronJobService;
+    httpService: HttpService;
 
   const signerMock = {
     address: MOCK_ADDRESS,
@@ -80,10 +76,6 @@ describe('WebhookService', () => {
         },
         { provide: ConfigService, useValue: mockConfigService },
         { provide: HttpService, useValue: createMock<HttpService>() },
-        {
-          provide: CronJobService,
-          useValue: createMock<CronJobService>(),
-        },
       ],
     }).compile();
 
@@ -91,7 +83,6 @@ describe('WebhookService', () => {
     webhookRepository = moduleRef.get(WebhookRepository);
     web3Service = moduleRef.get<Web3Service>(Web3Service);
     httpService = moduleRef.get<HttpService>(HttpService);
-    cronJobService = moduleRef.get<CronJobService>(CronJobService);
   });
 
   afterEach(() => {
@@ -289,132 +280,6 @@ describe('WebhookService', () => {
           event_type: webhookEntity.eventType,
         },
         { headers: { [HEADER_SIGNATURE_KEY]: expect.any(String) } },
-      );
-    });
-  });
-
-  describe('processPendingCronJob', () => {
-    let sendWebhookMock: any;
-    let cronJobEntityMock: Partial<CronJobEntity>;
-    let webhookEntity1: Partial<WebhookEntity>,
-      webhookEntity2: Partial<WebhookEntity>;
-
-    beforeEach(() => {
-      cronJobEntityMock = {
-        cronJobType: CronJobType.ProcessPendingWebhook,
-        startedAt: new Date(),
-      };
-
-      webhookEntity1 = {
-        id: 1,
-        chainId: ChainId.LOCALHOST,
-        escrowAddress: MOCK_ADDRESS,
-        status: WebhookStatus.PENDING,
-        waitUntil: new Date(),
-        retriesCount: 0,
-      };
-
-      webhookEntity2 = {
-        id: 2,
-        chainId: ChainId.LOCALHOST,
-        escrowAddress: MOCK_ADDRESS,
-        status: WebhookStatus.PENDING,
-        waitUntil: new Date(),
-        retriesCount: 0,
-      };
-
-      jest
-        .spyOn(webhookRepository, 'find')
-        .mockResolvedValue([webhookEntity1 as any, webhookEntity2 as any]);
-
-      sendWebhookMock = jest.spyOn(webhookService as any, 'sendWebhook');
-      sendWebhookMock.mockResolvedValue(true);
-
-      jest.spyOn(cronJobService, 'isCronJobRunning').mockResolvedValue(false);
-    });
-
-    afterEach(() => {
-      jest.restoreAllMocks();
-    });
-
-    it('should not run if cron job is already running', async () => {
-      jest
-        .spyOn(cronJobService, 'isCronJobRunning')
-        .mockResolvedValueOnce(true);
-
-      const startCronJobMock = jest.spyOn(cronJobService, 'startCronJob');
-
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(startCronJobMock).not.toHaveBeenCalled();
-    });
-
-    it('should create cron job entity to lock the process', async () => {
-      jest
-        .spyOn(cronJobService, 'startCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(cronJobService.startCronJob).toHaveBeenCalledWith(
-        CronJobType.ProcessPendingWebhook,
-      );
-    });
-
-    it('should send webhook for all of the pending webhooks', async () => {
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(sendWebhookMock).toHaveBeenCalledTimes(2);
-      expect(sendWebhookMock).toHaveBeenCalledWith(webhookEntity1);
-      expect(sendWebhookMock).toHaveBeenCalledWith(webhookEntity2);
-
-      expect(webhookRepository.updateOne).toHaveBeenCalledTimes(2);
-      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
-        { id: webhookEntity1.id },
-        { status: WebhookStatus.COMPLETED },
-      );
-      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
-        { id: webhookEntity2.id },
-        { status: WebhookStatus.COMPLETED },
-      );
-    });
-
-    it('should increase retriesCount by 1 if sending webhook fails', async () => {
-      sendWebhookMock.mockRejectedValueOnce(new Error());
-
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
-        { id: webhookEntity1.id },
-        {
-          retriesCount: 1,
-          waitUntil: expect.any(Date),
-        },
-      );
-    });
-
-    it('should mark webhook as failed if retriesCount exceeds threshold', async () => {
-      sendWebhookMock.mockRejectedValueOnce(new Error());
-
-      webhookEntity1.retriesCount = MOCK_MAX_RETRY_COUNT;
-
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(webhookRepository.updateOne).toHaveBeenCalledWith(
-        { id: webhookEntity1.id },
-        { status: WebhookStatus.FAILED },
-      );
-    });
-
-    it('should complete the cron job entity to unlock', async () => {
-      jest
-        .spyOn(cronJobService, 'completeCronJob')
-        .mockResolvedValueOnce(cronJobEntityMock as any);
-
-      await (webhookService as any).processPendingWebhooks();
-
-      expect(cronJobService.completeCronJob).toHaveBeenCalledWith(
-        cronJobEntityMock as any,
       );
     });
   });
