+++ packages/apps/job-launcher/server/vercel.json
@@ -17,27 +17,5 @@
       }
     }
   ],
-  "crons": [
-    {
-      "path": "/job/cron/create-escrow",
-      "schedule": "*/10 * * * *"
-    },
-    {
-      "path": "/job/cron/setup-escrow",
-      "schedule": "*/10 * * * *"
-    },
-    {
-      "path": "/job/cron/fund-escrow",
-      "schedule": "*/10 * * * *"
-    },
-    {
-      "path": "/job/cron/cancel",
-      "schedule": "*/10 * * * *"
-    },
-    {
-      "path": "/webhook/cron",
-      "schedule": "*/10 * * * *"
-    }
-  ],
   "ignoreCommand": "git diff HEAD^ HEAD --quiet ."
 }
